+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## [1.8.1] - 2021-11-29
+- Internal: rename SendAsync to AsyncSuspender
+- Internal: add missing field inside impl Debug
+- Add tokio_full.rs example.
+
 ## [1.8.0] - 2021-11-29
-- fix(Flower): polling mutex only if needed inside try_recv(|value| {...}, |result| {...}) which is introduced in version 1.0.0, now everyting's working as expected.
-- feat(FlowerHandle): send_async() support can be used from any async runtime.
-- feat(Flower): fn is_canceled() to check cancelation added.
+- Fix(Flower): polling mutex only if needed inside try_recv(|value| {...}, |result| {...}) which is introduced in version 1.0.0, now everyting's working as expected.
+- Feat(FlowerHandle): send_async() support can be used from any async runtime.
+- Feat(Flower): fn is_canceled() to check cancelation added.
 
 ## [1.0.0] - 2021-11-28
 - Improvement(Flower): instead of polling the Mutex over and over, poll the mutex inside fn try_recv(|value| {...}, |result| {...}) only if needed.